fix(unified-memory): spec-decoding correctness batch + band diagnostics
GPU-validated batch closing the spec×unified correctness campaign
plus the band-pressure hardening it rode on.

ROOT-CAUSE FIX (the headline): the spec-commit track-state scatter wrote
with UNTRANSLATED (virtual) mamba slot ids — set_mamba_track_indices_from_reqs
rebuilds batch.mamba_track_indices from allocator-minted ping-pong ids and
both spec commit callers (spec_utils, dflash_worker_v2) passed it raw into
update_mamba_state_after_mtp_verify, whose fused scatters address PHYSICAL
views. Silent wrong-slot writes once eviction/compaction breaks v2p≈identity:
radix-only, spec-only, overlap-amplified GSM8K collapse (-51pt worst,
accept pinned at 4.0). Fixed by translating at the commit seam (all three
algorithms). A GPU run reproduced 9/48 corrupt cells; the re-run validated
0/48, accept 1.56 healthy, penumbra gone.

Verify read-path hardening (DFLASH boot crashes):
- triton_backend: _translate_kv_indices_ragged — translate only the ragged
  prefix the fill kernel wrote (decode/verify/extend); over-allocated tails
  are uninitialized and device-assert in index_select.
- triton_backend: zeros-init verify kv_indices when the caller-provided
  seq_lens_sum is an UPPER BOUND (DFLASH host mirrors = prefix + block).
- triton_backend: clamp the target-verify custom_mask copy; hybrid backend
  forwards get/update_verify_buffers_to_fill_after_draft (EAGLE cg_on).
- memory_pool: page-major set_kv_buffer_prefix_valid implemented via
  valid-row select + native set_kv_buffer (was NotImplementedError).
- kv_cache_mixin: draft-worker virtual sizing gated by isinstance
  MultiEndedAllocator (duck-typed attr matched baseline SWA composite).
- hybrid backend: force Triton causal-conv under spec (intermediate windows).

DFLASH enablement: allow-list gate opened; unpin_spec_state_band public
helper; dflash commit wrapped in try/finally unpin (band unpinned on every exit).

Band pressure/accounting:
- SPEC_BAND_ALIGNMENT_MARGIN_SLOTS=3 reserve, charged in admission/decode
  budgets (schedule_policy offsets, check_decode_mem probe->place->retract);
  per-RUNNING-req band charge split from mamba_slot_full_token_cost.
- mamba prefetch capped by byte-budgeted remaining capacity + allocatable
  reqs (uncapped queue-sized prefetch claimed ~18GB intra-pass at mfs0.45).
- chunk-mode mamba sizing budget-capped (upstream reservation went negative
  at mfs0.45 -> boot failure); intermediate reservation sized by the
  REACHABLE capped request count, jointly with max_mamba_cache_size.
- band credit walk made compaction-mode-independent (eager double-count).
- urgent-flush mamba copy chunked per layer (13-15GiB transient OOM).
- i32->i64 scatter stride fix + production-stride regression test.

Diagnostics (env SGLANG_DEBUG_SPEC_BAND, off in production): import-time
liveness marker + three log-then-raise checks — band/mamba byte overlap at
place_band, verify translation freshness + band/read aliasing at metadata
build, commit-target liveness (incl. track lane) + pin-window at commit.

+++ python/sglang/srt/layers/attention/mamba/mamba_state_scatter_triton.py
@@ -183,8 +183,13 @@ def _fused_mamba_state_scatter_with_mask_kernel(
     # Load destination index
     dst_idx = tl.load(dst_indices_raw_ptr + pid_req).to(tl.int64)
 
-    # Source index is just the request index itself
-    src_idx = pid_req
+    # Source index is just the request index itself. MUST be int64 before it
+    # multiplies src_req_stride: the band-slice slot stride is ~1.4e8 ELEMENTS,
+    # so an i32 product wraps negative from src_idx >= 16 — an illegal address
+    # when the band sits low in the buffer (eval_228 eager+chunk crash) and
+    # SILENT state corruption when it sits high enough that the wrapped
+    # address stays inside the allocation (the eval_224 accept-len anomaly).
+    src_idx = pid_req.to(tl.int64)
 
     # Bounds check to avoid illegal memory access
     if not (
@@ -364,7 +369,10 @@ def _fused_conv_window_scatter_with_mask_kernel(
         return
 
     dst_idx = tl.load(dst_indices_raw_ptr + pid_req).to(tl.int64)
-    src_idx = pid_req
+    # int64 BEFORE the src_req_stride multiply — same i32-wrap hazard as the
+    # ssm kernel above (the deduped conv view's slot stride is ~2.8e8 elements,
+    # wrapping from src_idx >= 8).
+    src_idx = pid_req.to(tl.int64)
 
     if not (
         (dst_idx >= 0)
